se muestra imagén de usuario auth de google

+++ src/components/wall.js
@@ -1,4 +1,8 @@
-import { savePosts, getAllPosts, deletePost } from '../lib/configFirestore';
+import {
+  savePosts, getAllPosts, deletePost, getCurrentUser,
+} from '../lib/configFirestore';
+
+import { signOutGoogle } from '../lib/authUser';
 
 export const wall = (onNavigate) => {
   const wallDiv = document.createElement('div');
@@ -34,6 +38,8 @@ export const wall = (onNavigate) => {
   // textClose.textContent = 'Cerrar Sesión';
   postButton.textContent = 'Publicar';
 
+
+
   wallHeader.appendChild(imageLogoHeader);
   signOutButton.appendChild(imageClose);
   // signOutButton.appendChild(textClose);
@@ -48,18 +54,19 @@ export const wall = (onNavigate) => {
   postButton.addEventListener('click', () => {
     savePosts(textArea.value).then(() => {
       location.reload();
-      //alert('Publicación guardada con éxito');
+      // alert('Publicación guardada con éxito');
     }).catch(() => {
       alert('algo salió mal');
     });
   });
-  
+
   postContainer.innerHTML = ' ';
   getAllPosts().then((respuesta) => {
-    console.log(respuesta);
+    // console.log(respuesta);
     respuesta.forEach((element) => {
-      console.log(element.data());
+      // console.log(element.data());
       const dataPost = element.data();
+      console.log(dataPost);
       // se crean tarjetas de post
       const cardPost = document.createElement('div');
       cardPost.classList.add('divPostUsers');
@@ -68,45 +75,78 @@ export const wall = (onNavigate) => {
       const emailUser = document.createElement('p');
       emailUser.textContent = dataPost.email;
       emailUser.classList.add('emailUser');
+      const photoUserAuth = document.createElement('img');
+      photoUserAuth.src = dataPost.img;
+      
+      photoUserAuth.classList.add('photoUserAuth');
       const contentPost = document.createElement('div');
       contentPost.classList.add('contenPost');
 
       const postContentText = document.createElement('p');
       postContentText.textContent = dataPost.text;
-      const deletePostButton = document.createElement('button');
-      deletePostButton.classList.add('deletePostButton');
-      deletePostButton.textContent = 'Eliminar';
 
-      deletePostButton.addEventListener("click", () => {
-         
-        deletePost(element.id).then(() => {
-          console.log("elemento eliminado: ", element.id)
-          alert('Publicación eliminda');
-          location.reload();
-          //postContainer.innerHTML = " "
-        }).catch(() => {
-          alert('algo salió mal');
+      const currentUser = getCurrentUser();
+      console.log(currentUser);
+
+      if (element.data().email === currentUser) {
+        const deletePostButton = document.createElement('button');
+        deletePostButton.classList.add('deletePostButton');
+        deletePostButton.textContent = 'Eliminar';
+        cardDiv.appendChild(deletePostButton);
+
+        deletePostButton.addEventListener('click', () => {
+          deletePost(element.id).then(() => {
+            console.log('elemento eliminado: ', element.id);
+            alert('Publicación eliminda');
+            location.reload();
+          // postContainer.innerHTML = " "
+          }).catch(() => {
+            alert('algo salió mal');
+          });
+
+          // console.log(deletePost());
+
+        // alert(element.id);
         });
-        
-       // console.log(deletePost());
 
-        //alert(element.id);
-      })
-
-       
+      }
+      cardDiv.appendChild(photoUserAuth);
       cardDiv.appendChild(emailUser);
       contentPost.appendChild(postContentText);
       cardPost.appendChild(cardDiv);
       cardPost.appendChild(contentPost);
-      postContainer.appendChild(cardPost); 
-      cardDiv.appendChild(deletePostButton);
-      
+      postContainer.appendChild(cardPost);
     });
   });
 
+
+
+  // Función para eliminar cookies manualmente
+function deleteAllCookies() {
+  var cookies = document.cookie.split(";");
+  for (var i = 0; i < cookies.length; i++) {
+    var cookie = cookies[i];
+    var eqPos = cookie.indexOf("=");
+    var name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie;
+    document.cookie = name + "=;expires=Thu, 01 Jan 1970 00:00:00 GMT";
+  }
+}
+
+  //cerrar sesión y eliminar cookies manualmente
   signOutButton.addEventListener('click', () => {
-    onNavigate('/');
+    signOutGoogle().then(() => {
+      deleteAllCookies(); // Eliminar cookies manualmente
+      alert("sesión cerrada")
+      onNavigate('/');
+    }).catch((error) => {
+      alert("error:" + error)
+      
+    });
+
+   
   });
+
+  
 
   return wallDiv;
 };
